requirements/minimal.txt
torch==1.8.1+cu111 -f https://download.pytorch.org/whl/cu111/torch_stable.html
torchvision==0.9.1+cu111 -f https://download.pytorch.org/whl/cu111/torch_stable.html
tensorboard==2.7.0
torch-tb-profiler==0.3.1
ninja==1.10.2
numpy==1.21.5
scipy==1.7.3
scikit-learn==1.0.2
scikit-video==1.1.11
pillow==9.0.0
opencv-python-headless==4.5.5.62
requests
bs4
tqdm
rich
click
cloup
psutil
easydict
lmdb
matplotlib
